goals added

+++ app/src/main/java/com/budget/tracker/requests/AddIncomeRequest.kt
@@ -1,3 +1,3 @@
 package com.budget.tracker.requests
 
-data class AddIncomeRequest (val categoryId: Int, val amount:Int, val currencyId:Int)
+data class AddIncomeRequest (val categoryId: Int, val amount:Double, val currencyId:Int)
